+++ docs/pages/components/button/examples/ExSizes.vue
@@ -1,0 +1,10 @@
+<template>
+    <section>
+        <div class="buttons">
+            <b-button size="is-small">Small</b-button>
+            <b-button>Default</b-button>
+            <b-button size="is-medium">Medium</b-button>
+            <b-button size="is-large">Large</b-button>
+        </div>
+    </section>
+</template>