+++ docs/pages/components/button/examples/ExTags.vue
@@ -1,0 +1,17 @@
+<template>
+    <section>
+        <div class="buttons">
+            <b-button>Button</b-button>
+
+            <b-button tag="a"
+                href="https://buefy.org"
+                target="_blank">
+                Anchor
+            </b-button>
+
+            <b-button tag="input"
+                native-type="submit"
+                value="Submit input" />
+        </div>
+    </section>
+</template>