+++ docs/pages/components/button/examples/ExTypesStates.vue
@@ -1,0 +1,53 @@
+<template>
+    <section>
+        
+        <div class="buttons">
+            <b-button type="is-primary">Primary</b-button>
+
+            <b-button type="is-success">Success</b-button>
+
+            <b-button type="is-danger">Danger</b-button>
+ 
+            <b-button type="is-warning">Warning</b-button>
+        
+            <b-button type="is-info">Info</b-button>
+        
+            <b-button type="is-link">Link</b-button>
+
+            <b-button type="is-light">Light</b-button>
+
+            <b-button type="is-dark">Dark</b-button>
+
+            <b-button type="is-text">Text</b-button>
+        </div>
+
+        <div class="buttons">
+            <b-button disabled>Disabled</b-button>
+        
+            <b-button loading>Loading</b-button>
+        
+            <b-button rounded>Rounded</b-button>
+        </div>
+
+        <div class="buttons">
+            <b-button type="is-primary" outlined>Outlined</b-button>
+            <b-button type="is-success" outlined>Outlined</b-button>
+            <b-button type="is-danger" outlined>Outlined</b-button>
+            <b-button type="is-warning" outlined>Outlined</b-button>
+        </div>
+
+        <div class="notification is-primary">
+            <b-button type="is-primary"
+                inverted>
+                Inverted
+            </b-button>
+
+            <b-button type="is-primary"
+                inverted
+                outlined>
+                Invert Outlined
+            </b-button>
+        </div>
+
+    </section>
+</template>