+++ src/components/index.js
@@ -1,4 +1,5 @@
 import Autocomplete from './autocomplete'
+import Button from './button'
 import Checkbox from './checkbox'
 import Collapse from './collapse'
 import Clockpicker from './clockpicker'
@@ -28,6 +29,7 @@ import Upload from './upload'
 
 export {
     Autocomplete,
+    Button,
     Checkbox,
     Clockpicker,
     Collapse,